Finish draft version to compute thermal Lyapunov spectrum using Haar random states.
One thing I think we need to improve is actually for system with large state space : ~ 33 dof or 36 dof, best way to do simulation is :
when evaluating a_{i}(t): we should first evolve forward in time and then evolve backward in time .
Previously we insert complete  basis set , this will be extremely expensive when dealing with high dimensional state space.

With evolving forward and backward in time, we do not suffer from dimensional curse.